feat: turns simplejwt as optional library and test it failure
