--- rest_framework_supertest/authentication/simple_jwt/__init__.py
@@ -1,21 +1,0 @@
-from .base import SimpleJWTAuthentication
-from .errors import (
-    NO_ACTIVE_ACCOUNT,
-    TOKEN_NO_RECOGNIZABLE_USER_ID,
-    TOKEN_NOT_VALID_FOR_ANY_TOKEN_TYPE,
-    TWO_AUTORIZATION_PARTS,
-    USER_IS_INACTIVE,
-    USER_NOT_FOUND,
-    USER_PASSWORD_CHANGED,
-)
-
-__all__ = [
-    'SimpleJWTAuthentication',
-    'TWO_AUTORIZATION_PARTS',
-    'NO_ACTIVE_ACCOUNT',
-    'TOKEN_NOT_VALID_FOR_ANY_TOKEN_TYPE',
-    'TOKEN_NO_RECOGNIZABLE_USER_ID',
-    'USER_NOT_FOUND',
-    'USER_IS_INACTIVE',
-    'USER_PASSWORD_CHANGED',
-]
